proyecto/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mysql-connector-python==8.2.0
bcrypt==4.1.2
PyJWT==2.8.0
python-multipart==0.0.6
pydantic==2.6.4
passlib[bcrypt]==1.7.4